(PC-29518)[PRO] refactor: remove upsertStocksAdapter
