+++ pro/src/screens/IndividualOffer/StocksEventEdition/serializers.ts
